Refactor: Fixed to divide into compornents
